broke StandardLibrary.NativeMethods into multiple partial files
